==> 1. Load and unzip emails.ipynb <==
# This script loads emails tar.gz file and unpacks into appropriate directories structure

### Download emails


from urllib import urlretrieve
url='http://www.cs.cmu.edu/~./enron/enron_mail_20150507.tgz'
dst='enron_mail_20150507.tgz'
urlretrieve(url,dst)

### Extract file into directories structure


import tarfile
tar = tarfile.open("enron_mail_20150507fast.tgz")
tar.extractall(path="Enron2")
tar.close()
